// rentautosFront/src/vendedores/vendedor.model.d.ts
export interface vendedorcreacionDTO{
    
    nombre:string;    
    cedula:string;
    celular:string;
    direccion:string;
    email:string;
    fechaNacimiento?:Date;
    foto:string;
   
}

export interface vendedorDTO{
    id:number;
    nombre:string;    
    cedula:string;
    celular:string;
    direccion:string;
    email:string;
    fechaNacimiento?:Date;
    foto:string;
}
    